Crud operation done

+++ src/components/CoffeeCard.jsx
@@ -1,26 +1,40 @@
-import Swal from 'sweetalert2'
+import PropTypes from 'prop-types';
+import { Link } from 'react-router-dom';
+import Swal from 'sweetalert2';
+
 const CoffeeCard = ({ coffee }) => {
-  const { name,_id, photoUrl, quantity, taste } = coffee;
+  const { name, _id, photoUrl, quantity, taste } = coffee;
 
+  const handleDelete = (_id) => 
 
-    const handleDelete =  (_id) => { 
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to delete this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (result.isConfirmed) {
         fetch(`http://localhost:5000/coffee/${_id}`, {
-            method: "DELETE",
+          method: "DELETE",
         })
-        .then((res) => res.json())
-        .then((data) => {
+          .then((res) => res.json())
+          .then((data) => {
             console.log(data);
             if (data.deletedCount > 0) {
-                Swal.fire({
-                    title: 'Success',
-                    text: 'Coffee Deleted successfully',
-                    icon: 'success',
-                    confirmButtonText: 'Okh'
-                })
+              Swal.fire({
+                title: 'Success',
+                text: 'Coffee Deleted successfully',
+                icon: 'success',
+                confirmButtonText: 'Okh'
+              });
             }
-        });
-        }
-        
+          });
+      }
+    });
+
   return (
     <div>
       <div className="card card-side bg-base-100 shadow-xl">
@@ -40,7 +54,9 @@ const CoffeeCard = ({ coffee }) => {
           <div className="card-actions justify-end">
             <div className="join join-vertical space-y-4">
               <button className="btn btn-success">View</button>
+              <Link to={`/updateCoffee/${_id}`}>
               <button className="btn btn-info">Edit</button>
+              </Link>
               <button onClick={()=> handleDelete(_id)} className="btn btn-error">X</button>
             </div>
           </div>
@@ -48,6 +64,16 @@ const CoffeeCard = ({ coffee }) => {
       </div>
     </div>
   );
+};
+
+CoffeeCard.propTypes = {
+  coffee: PropTypes.shape({
+    name: PropTypes.string.isRequired,
+    _id: PropTypes.string.isRequired,
+    photoUrl: PropTypes.string.isRequired,
+    quantity: PropTypes.number.isRequired,
+    taste: PropTypes.string.isRequired,
+  }).isRequired,
 };
 
 export default CoffeeCard;
